requirements.txt
lxml
mechanize
python-amazon-product-api
simplejson
PIL
BeautifulSoup
pycurl
pylast
tweepy
